WIP(arch-refactor): S1-S3a 完了チェックポイント + S3b-1 部分適用
S1 common 基盤 (protocol 4分割/fingerprint/device_id v29/blob-less wire、330 tests)
S2a daw_audio (有界dispatch+quarantine/render統一/adopt_state_from/RtBundle/cpal 0.17、79 tests)
S2b daw_plugin_host (InstanceRecord device_id 一本化/CLAP ext 完備/ProcessScaffold/split-half、35 tests)
S3a daw_gui 新protocol機械適応 (fingerprint検証/writer死検知/respawn rate/世代guard、280 tests)
S7 再発防止 (CLAUDE.md 不変条件/arch-lint/guards 7則/arch-review skill)
S6 DESIGN.md 現行化 / docs/plan_arch_refactor.md

注意: daw_gui は S3b-1 (state/ 分割) の agent stall により半変換状態 (176 errors)。
このコミットは復旧作業の防波堤 (検証済み S1-S3a を diff 可能に保全する)。

+++ common/src/worker_bridge.rs
@@ -4,34 +4,34 @@
 //! The two processes each spawn N worker threads in 1:1 pairs. When audio
 //! engine `worker[i]` wants `plugin_host worker[i]` to call
 //! `plugin.process()` for a particular plugin instance, it writes the
-//! plugin id into `worker_task[i]` (Release), then signals `worker_wake[i]`
-//! (a Win32 named event). The plugin-host worker reads the slot (Acquire)
-//! after waking, runs `process()`, and signals `worker_done[i]`.
+//! stable **device id** (v29, `PluginInstance::id`) into `worker_task[i]`
+//! (Release), then signals `worker_wake[i]` (a Win32 named event). The
+//! plugin-host worker reads the slot (Acquire) after waking, runs
+//! `process()`, and signals `worker_done[i]`.
 //!
 //! Only one shmem instance exists for the whole daw_01 session — its name
 //! is fixed (see `plugin_ref::worker_bridge_shmem_id`).
 
-use std::sync::atomic::AtomicU32;
+use std::sync::atomic::AtomicU64;
 
 /// Hard cap on workers. CPU core counts above this are extremely rare for
-/// audio workloads (2026: typical desktop is 4–24 cores). Sized so the
-/// whole struct stays under one cache line on most machines.
+/// audio workloads (2026: typical desktop is 4–24 cores).
 pub const MAX_WORKERS: usize = 32;
 
 #[repr(C)]
 pub struct WorkerBridge {
-    /// `worker_task[i]` is the plugin id audio-engine `worker[i]` is
-    /// asking plugin-host `worker[i]` to process this dispatch.
-    /// `u32::MAX` means "idle, ignore" (used during shutdown wake).
-    pub worker_task: [AtomicU32; MAX_WORKERS],
+    /// `worker_task[i]` is the stable device id audio-engine `worker[i]`
+    /// is asking plugin-host `worker[i]` to process this dispatch.
+    /// `u64::MAX` means "idle, ignore" (used during shutdown wake).
+    pub worker_task: [AtomicU64; MAX_WORKERS],
 }
 
 impl WorkerBridge {
-    pub const IDLE: u32 = u32::MAX;
+    pub const IDLE: u64 = u64::MAX;
 
     pub fn zeroed() -> Self {
         Self {
-            worker_task: [const { AtomicU32::new(Self::IDLE) }; MAX_WORKERS],
+            worker_task: [const { AtomicU64::new(Self::IDLE) }; MAX_WORKERS],
         }
     }
 }
@@ -53,7 +53,7 @@ mod shmem_handle {
             // The shmem view is backed by `MapViewOfFile`, which returns a
             // pointer aligned to at least 64 KiB (the system allocation
             // granularity). That trivially satisfies `WorkerBridge`'s
-            // alignment (== `AtomicU32` align == 4).
+            // alignment (== `AtomicU64` align == 8).
             debug_assert!(
                 (shmem.as_ptr() as usize).is_multiple_of(std::mem::align_of::<WorkerBridge>()),
                 "worker_bridge shmem pointer is not WorkerBridge-aligned"
@@ -75,7 +75,7 @@ mod shmem_handle {
         pub fn open(os_id: &str) -> Result<Self> {
             let shmem = NamedShmem::open(os_id, std::mem::size_of::<WorkerBridge>())?;
             // `MapViewOfFile` aligns the view to the 64 KiB system allocation
-            // granularity, which covers `WorkerBridge`'s 4-byte alignment.
+            // granularity, which covers `WorkerBridge`'s 8-byte alignment.
             debug_assert!(
                 (shmem.as_ptr() as usize).is_multiple_of(std::mem::align_of::<WorkerBridge>()),
                 "worker_bridge shmem pointer is not WorkerBridge-aligned"
@@ -86,8 +86,8 @@ mod shmem_handle {
         pub fn bridge(&self) -> &WorkerBridge {
             // SAFETY: the mapping is at least `size_of::<WorkerBridge>()` bytes
             // (checked in `create`/`open`) and the `MapViewOfFile` pointer is
-            // 64 KiB-aligned, satisfying `WorkerBridge`'s 4-byte alignment.
-            // `WorkerBridge` is `#[repr(C)]` and holds only `AtomicU32`s, which
+            // 64 KiB-aligned, satisfying `WorkerBridge`'s 8-byte alignment.
+            // `WorkerBridge` is `#[repr(C)]` and holds only `AtomicU64`s, which
             // are valid for any bit pattern, so the mapped bytes are always a
             // valid `WorkerBridge`. Concurrent access from the peer process is
             // sound because all fields are atomics.
